requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
backoff==2.2.1
black==25.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
chromadb==1.3.0
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
langfuse==2.60.9
langfuse-langchain>=2.0.0
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-openai==1.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.4
mistune-telegram==0.5.0
mypy_extensions==1.1.0
openai==1.109.1 # langchain-openai 1.0.1 depends on openai<3.0.0 and >=1.109.1
orjson==3.11.3
packaging==24.2
pathspec==0.12.1
platformdirs==4.5.0
postgrest==1.1.1
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.2.1
python-multipart==0.0.20
pytokens==0.2.0
PyYAML==6.0.2
realtime==2.7.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
tenacity==9.1.2
tqdm==4.67.1
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3